Aggregate equal stackable subscriptions
This is useful for OEM subscriptions (multiple subscriptions with
quantity == 1) and systems that require more than 1 "piece" of
subscription to be covered (e.g. systems with a lot of CPUs).

As the current implementation of matcher can't assign multiple
subscriptions that cover same product (that is forbidden by the
conflicts in the move definition), this change will aggregate equal
subscription into a single one (and adjusting its quantity).